Refactor string formatting functions into separate module
