+++ pyproprop/__init__.py
@@ -1,2 +1,5 @@
+from .format_str_case import format_str_case
+
 from .named_iterable import named_iterable
+
 from .processed_property import processed_property
